micro/service/poc_scanner/requirement.txt
python-consul==1.1.0